fix: HuggingFace backend hardware-specific compute dtype fix for GTX 2070
## Problem
HuggingFace backend was using torch.bfloat16 as compute dtype, which causes:
- Silent numerical corruption on GTX 2070 (Turing architecture CC 7.5)
- bfloat16 only supported on Ampere (RTX 30 series) and newer
- Resulted in gibberish output, repetition, and looping audio

## Solution
Changed bnb_4bit_compute_dtype from torch.bfloat16 to torch.float16:
- GTX 2070 (Turing) has native FP16 hardware acceleration (Tensor Cores)
- Enables fast, correct numerical computation
- Fixes quality issues: no more gibberish, repetition, or looping

## Optimal Settings (HuggingFace Backend)
- temperature: 0.43
- top_p: 0.90
- max_tokens: 2500
- repetition_penalty: 1.10

## Emotion Tag Format (Important)
Use single tags only, NOT opening/closing pairs:
✅ Correct: "The forest was quiet. <whisper> Something watched from shadows."
❌ Incorrect: "The forest was quiet. <whisper>Something watched</whisper>"

## Testing
Created test_hf_backend_quick.py with validation showing:
- Clean audio generation (no gibberish/NaN)
- Proper RMS levels (0.058-0.117)
- No clipping (peaks 0.506-0.843)
- Natural breathing and prosody with emotion tags

Reference: LLM Performance Optimization Report identified hardware mismatch
as critical issue causing silent numerical corruption on Turing GPUs.

+++ core/tts_maya1_hf.py
@@ -36,7 +36,7 @@ def _ensure_models(model_path: str):
                 load_in_4bit=True,
                 bnb_4bit_use_double_quant=True,
                 bnb_4bit_quant_type="nf4",
-                bnb_4bit_compute_dtype=torch.bfloat16,
+                bnb_4bit_compute_dtype=torch.float16,  # GTX 2070 (Turing CC 7.5) has native FP16 support, not bfloat16
             )
 
             # Load model with 4-bit quantization on GPU
